pyproject.toml
[project]
name = "user-db-indexer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "pyjwt>=2.10.1",
    "uvicorn>=0.40.0",
]
